refactor: Comment 변환 로직을 CommentMapper로 통합 #130
- mapper 코드 이동 및 메서드 추가
