+++ src/main/java/com/knu/sosuso/capstone/domain/video/service/PopularVideoService.java
@@ -9,7 +9,7 @@ import com.knu.sosuso.capstone.domain.video.repository.VideoStatsRepository;
 import com.knu.sosuso.capstone.domain.video.service.popularity.PopularityCalculator;
 import com.knu.sosuso.capstone.global.exception.BusinessException;
 import com.knu.sosuso.capstone.global.exception.error.TrendingError;
-import com.knu.sosuso.capstone.global.service.ResponseMappingService;
+import com.knu.sosuso.capstone.global.service.mapper.ResponseMappingService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.annotation.Cacheable;
