+++ src/main/java/com/knu/sosuso/capstone/global/service/mapper/CommentMapper.java
@@ -1,4 +1,4 @@
-package com.knu.sosuso.capstone.domain.comment.service;
+package com.knu.sosuso.capstone.global.service.mapper;
 
 import com.knu.sosuso.capstone.domain.ai.dto.AIAnalysisResponse;
 import com.knu.sosuso.capstone.domain.comment.dto.CommentDto;
@@ -116,6 +116,86 @@ public class CommentMapper {
                         return toDto(commentData);
                     }
                 })
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * Comment → CommentDto 변환 (TOP 5 댓글용)
+     * hasReplies를 강제로 false로 설정
+     */
+    public CommentDto toTopCommentDto(Comment comment) {
+        return new CommentDto(
+                comment.getApiCommentId(),
+                comment.getWriter(),
+                comment.getCommentContent(),
+                comment.getLikeCount(),
+                comment.getSentimentType() != null ?
+                        comment.getSentimentType().name().toUpperCase() : null,
+                comment.getWrittenAt(),
+                false,  // TOP 5 댓글은 항상 false
+                comment.getDetailSentiments() != null ?
+                        comment.getDetailSentiments().stream()
+                                .map(Enum::name)
+                                .collect(Collectors.toList()) :
+                        new ArrayList<>()
+        );
+    }
+
+    /**
+     * CommentApiResponse.CommentData → CommentDto 변환 (TOP 5 댓글용)
+     * hasReplies를 강제로 false로 설정
+     */
+    public CommentDto toTopCommentDto(CommentApiResponse.CommentData commentData,
+                                      AIAnalysisResponse analysisResponse) {
+
+        String sentiment = null;
+        List<String> detailSentiments = new ArrayList<>();
+
+        if (analysisResponse != null) {
+            CommentSentimentDetail sentimentDetail = analysisResponse.sentimentComments().stream()
+                    .filter(detail -> detail.apiCommentId().equals(commentData.id()))
+                    .findFirst()
+                    .orElse(null);
+
+            if (sentimentDetail != null) {
+                sentiment = sentimentDetail.sentimentType().name().toUpperCase();
+                detailSentiments = sentimentDetail.detailSentimentTypes().stream()
+                        .map(Enum::name)
+                        .collect(Collectors.toList());
+            }
+        }
+
+        return new CommentDto(
+                commentData.id(),
+                commentData.authorName(),
+                commentData.commentText(),
+                commentData.likeCount(),
+                sentiment,
+                commentData.publishedAt(),
+                false,  // TOP 5 댓글은 항상 false
+                detailSentiments
+        );
+    }
+
+    /**
+     * DB에서 가져온 TOP 5 댓글 리스트 변환
+     */
+    public List<CommentDto> toTopCommentDtoList(List<Comment> comments) {
+        return comments.stream()
+                .map(this::toTopCommentDto)
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * YouTube API에서 가져온 TOP 5 댓글 리스트 변환
+     */
+    public List<CommentDto> toTopCommentDtoList(List<CommentApiResponse.CommentData> commentDataList,
+                                                AIAnalysisResponse analysisResponse,
+                                                int limit) {
+        return commentDataList.stream()
+                .sorted((c1, c2) -> Integer.compare(c2.likeCount(), c1.likeCount()))
+                .limit(limit)
+                .map(commentData -> toTopCommentDto(commentData, analysisResponse))
                 .collect(Collectors.toList());
     }
 }
